Ajustes de transição
