+++ MarvelApp/app/src/main/java/com/example/marvelapp/alterarsenha/view/AlterarSenhaActivity.kt
@@ -1,36 +1,35 @@
-package com.example.marvelapp.cadastro
+package com.example.marvelapp.alterarsenha.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.widget.Toast
+import android.widget.Toolbar
 import androidx.appcompat.app.AlertDialog
 import com.example.marvelapp.R
-import com.example.marvelapp.home.view.HomeActivity
-import com.example.marvelapp.login.LoginActivity
-import com.google.android.material.appbar.MaterialToolbar
+import com.example.marvelapp.login.view.LoginActivity
 import com.google.android.material.button.MaterialButton
 import kotlinx.android.synthetic.main.dialog_confirmacao.view.*
 
-class CadastroActivity : AppCompatActivity() {
+class AlterarSenhaActivity : AppCompatActivity() {
 
     private var alertDialog: AlertDialog? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_cadastro)
+        setContentView(R.layout.activity_alterar_senha)
 
-        val toolbarCadastro = findViewById<MaterialToolbar>(R.id.toolbarCadastro)
-        toolbarCadastro.setOnClickListener {
+        val toolbar = findViewById<Toolbar>(R.id.toolbarAlterarSenha)
+        toolbar.setNavigationOnClickListener {
             showDialog()
         }
 
-        val btnCadastro = findViewById<MaterialButton>(R.id.btnCadastro)
-        btnCadastro.setOnClickListener {
-            Toast.makeText(this, "Campos vazios", Toast.LENGTH_SHORT).show()
+        val btnAlterarSenha = findViewById<MaterialButton>(R.id.btnSalvarSenha)
+        btnAlterarSenha.setOnClickListener {
 
-            val intent = Intent(this@CadastroActivity, HomeActivity::class.java)
+            Toast.makeText(this, "Senha alterada com sucesso", Toast.LENGTH_SHORT).show()
+            val intent = Intent(this@AlterarSenhaActivity, LoginActivity::class.java)
             startActivity(intent)
             finish()
         }
@@ -38,20 +37,21 @@ class CadastroActivity : AppCompatActivity() {
 
     private fun showDialog() {
 
-        val dialogBuilder = AlertDialog.Builder(this@CadastroActivity)
+        val dialogBuilder = AlertDialog.Builder(this@AlterarSenhaActivity)
         val dialogView = LayoutInflater.from(this).inflate(R.layout.dialog_confirmacao, null, false)
         dialogBuilder.setView(dialogView)
 
         dialogView.btnContinuarEditando.setOnClickListener { alertDialog?.dismiss() }
 
-        dialogView.btnConfirmar.setOnClickListener {
-            val intent = Intent(this, LoginActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
+        dialogView.btnConfirmar.setOnClickListener { finish() }
 
         alertDialog = dialogBuilder.create()
         alertDialog?.window!!.setBackgroundDrawableResource(android.R.color.transparent)
         alertDialog?.show()
     }
+
+    override fun onBackPressed() {
+        showDialog()
+    }
+
 }
