+++ MarvelApp/app/src/main/java/com/example/marvelapp/cadastro/view/CadastroActivity.kt
@@ -1,38 +1,36 @@
-package com.example.marvelapp.alterarsenha
+package com.example.marvelapp.cadastro.view
 
-import android.app.Dialog
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.Window
 import android.widget.Toast
-import android.widget.Toolbar
 import androidx.appcompat.app.AlertDialog
 import com.example.marvelapp.R
-import com.example.marvelapp.login.LoginActivity
-import com.example.marvelapp.perfil.PerfilFragment
+import com.example.marvelapp.home.view.HomeActivity
+import com.example.marvelapp.login.view.LoginActivity
+import com.google.android.material.appbar.MaterialToolbar
 import com.google.android.material.button.MaterialButton
 import kotlinx.android.synthetic.main.dialog_confirmacao.view.*
 
-class AlterarSenhaActivity : AppCompatActivity() {
+class CadastroActivity : AppCompatActivity() {
 
     private var alertDialog: AlertDialog? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_alterar_senha)
+        setContentView(R.layout.activity_cadastro)
 
-        val toolbar = findViewById<Toolbar>(R.id.toolbarAlterarSenha)
-        toolbar.setNavigationOnClickListener {
+        val toolbarCadastro = findViewById<MaterialToolbar>(R.id.toolbarCadastro)
+        toolbarCadastro.setOnClickListener {
             showDialog()
         }
 
-        val btnAlterarSenha = findViewById<MaterialButton>(R.id.btnSalvarSenha)
-        btnAlterarSenha.setOnClickListener {
+        val btnCadastro = findViewById<MaterialButton>(R.id.btnCadastro)
+        btnCadastro.setOnClickListener {
+            Toast.makeText(this, "Campos vazios", Toast.LENGTH_SHORT).show()
 
-            Toast.makeText(this, "Senha alterada com sucesso", Toast.LENGTH_SHORT).show()
-            val intent = Intent(this@AlterarSenhaActivity, LoginActivity::class.java)
+            val intent = Intent(this@CadastroActivity, HomeActivity::class.java)
             startActivity(intent)
             finish()
         }
@@ -40,17 +38,24 @@ class AlterarSenhaActivity : AppCompatActivity() {
 
     private fun showDialog() {
 
-        val dialogBuilder = AlertDialog.Builder(this@AlterarSenhaActivity)
+        val dialogBuilder = AlertDialog.Builder(this@CadastroActivity)
         val dialogView = LayoutInflater.from(this).inflate(R.layout.dialog_confirmacao, null, false)
         dialogBuilder.setView(dialogView)
 
         dialogView.btnContinuarEditando.setOnClickListener { alertDialog?.dismiss() }
 
-        dialogView.btnConfirmar.setOnClickListener { finish() }
+        dialogView.btnConfirmar.setOnClickListener {
+            val intent = Intent(this, LoginActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
 
         alertDialog = dialogBuilder.create()
         alertDialog?.window!!.setBackgroundDrawableResource(android.R.color.transparent)
         alertDialog?.show()
     }
 
+    override fun onBackPressed() {
+        showDialog()
+    }
 }
