+++ MarvelApp/app/src/main/res/layout-v23/activity_alterar_senha.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/img_background"
     android:orientation="vertical"
-    tools:context=".alterarsenha.AlterarSenhaActivity">
+    tools:context=".alterarsenha.view.AlterarSenhaActivity">
 
     <Toolbar
         android:id="@+id/toolbarAlterarSenha"
