+++ MarvelApp/app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_app"
-    tools:context=".splashscreen.SplashScreenActivity">
+    tools:context=".splashscreen.view.SplashScreenActivity">
 
     <LinearLayout
         android:id="@+id/imgSplash"
