+++ MarvelApp/app/src/main/res/layout/item_card_favoritos.xml
@@ -41,7 +41,7 @@
                 android:id="@+id/txtNamePersongemFavoritado"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
-                android:layout_margin="@dimen/dim_10_dp"
+                android:layout_marginStart="@dimen/dim_10_dp"
                 android:layout_weight="1"
                 android:maxLines="2"
                 android:layout_gravity="center"
